Create README.md
This Website was submitted to Grand Circus Bootcamp in Detroit, MI as part of an application for participation in Front-End Web Development.
I went to college at Central Michigan University, and graduated with a Double Major in Marketing and Logistics Management.
Straight out of college, I started my career as a Marketing Consultant for a Regional Franchisor in the Food Industry.
At first, I loved it. I really enjoyed the people I worked, and still talk to some today. 
Overtime, I was able to see the direction of my career path, and realized that it is not something I wanted to commit my life too.
Setting price points and depth of understanding market you're in, amongst other conecpts, I found my passion fading.
I have always had a great visual memory. I have work with photoshop for 10+ years. 
First it started as a hobby when I was young, and later I was able to use that hobby in a professional capacity.
I have always had great problem solving mindset. Any puzzle/problem solving challenge, I have always enjoyed. 
Rubix Cubes, 3-D Puzzles, any game or challenge I could get my hand on. So I focused on that, what can I do that is aligned with that.
Shortly after that, I was working with 100+ Excel Spreadsheets. I would open each spreadsheet to enter the data as a report in another excel sheet.
After a short time googling, I found a way to link the sheets using a concat function and the file path.
I knew right then, that I want to do something like this for the rest of my life. I felt so accomplished, saved hours of worked with simple function and path.
I started researching and realized that is eactly what a Developer does. 
I started learning HTML, CSS and JavaScript. I found Grand Circus Bootcamps in Detroit, and thus this site was born.